+++ packages/graphql/config/environment.ts
@@ -1,3 +1,4 @@
+import { getExpoConstant } from '@baseapp-frontend/utils'
 import { baseAppFetch } from '@baseapp-frontend/utils/functions/fetch/baseAppFetch'
 import { getToken } from '@baseapp-frontend/utils/functions/token/getToken'
 
@@ -64,9 +65,10 @@ export async function httpFetch(
   uploadables?: UploadableMap | null,
 ): Promise<GraphQLResponse> {
   const fetchOptions = getFetchOptions({ request, variables, uploadables })
+  const EXPO_PUBLIC_RELAY_ENDPOINT = getExpoConstant('EXPO_PUBLIC_RELAY_ENDPOINT')
+
   const response = await baseAppFetch('', {
-    baseUrl: (process.env.NEXT_PUBLIC_RELAY_ENDPOINT ||
-      process.env.EXPO_PUBLIC_RELAY_ENDPOINT) as string,
+    baseUrl: (process.env.NEXT_PUBLIC_RELAY_ENDPOINT ?? EXPO_PUBLIC_RELAY_ENDPOINT) as string,
     decamelizeRequestBodyKeys: false,
     decamelizeRequestParamsKeys: false,
     camelizeResponseDataKeys: false,
@@ -89,9 +91,9 @@ export async function httpFetch(
   return response
 }
 
+const EXPO_PUBLIC_WS_RELAY_ENDPOINT = getExpoConstant('EXPO_PUBLIC_WS_RELAY_ENDPOINT')
 const wsClient = createClient({
-  url: (process.env.NEXT_PUBLIC_WS_RELAY_ENDPOINT ||
-    process.env.EXPO_PUBLIC_WS_RELAY_ENDPOINT) as string,
+  url: (process.env.NEXT_PUBLIC_WS_RELAY_ENDPOINT ?? EXPO_PUBLIC_WS_RELAY_ENDPOINT) as string,
   connectionParams: () => {
     const Authorization = getToken()
     if (!Authorization) return {}