+++ packages/utils/functions/os/index.ts
@@ -1,0 +1,7 @@
+import { Platform } from 'react-native'
+
+export const isMobilePlatform = () => {
+  if (!Platform) return false
+
+  return Platform.OS === 'ios' || Platform.OS === 'android'
+}