+++ packages/utils/functions/token/getTokenAsync/index.ts
@@ -3,12 +3,13 @@ import { getItemAsync } from 'expo-secure-store'
 import { ACCESS_KEY_NAME } from '../../../constants/jwt'
 import type { ServerSideRenderingOption } from '../../../types/server'
 import { getCookieAsync } from '../../cookie'
+import { isMobilePlatform } from '../../os'
 
 export const getTokenAsync = async (
   key = ACCESS_KEY_NAME,
   { noSSR = false }: ServerSideRenderingOption = {},
 ) => {
-  if (process.env.EXPO_PUBLIC_PLATFORM === 'mobile') {
+  if (isMobilePlatform()) {
     const token = await getItemAsync(key)
     return token
   }